push: Refactor Action serde code and move to submodule
